+++ dgraphclient/dgraph.go
@@ -2,7 +2,7 @@
  * Copyright 2024 Hypermode, Inc.
  */
 
-package dqlclient
+package dgraphclient
 
 import (
 	"context"
@@ -101,6 +101,85 @@ func (dc *dgraphConnector) executeMutations(ctx context.Context, setMutations, d
 
 	resp, err := tx.Do(ctx, req)
 	return resp.Uids, err
+}
+
+func (dc *dgraphConnector) execute(ctx context.Context, req Request) (Response, error) {
+	if len(req.Mutations) == 0 {
+		if req.Query.Query == "" {
+			return Response{}, nil
+		}
+
+		tx := dc.dgClient.NewReadOnlyTxn()
+		defer func() {
+			if err := tx.Discard(ctx); err != nil {
+				logger.Warn(ctx).Err(err).Msg("Error discarding transaction.")
+				return
+			}
+
+		}()
+
+		dgoReq := &api.Request{Query: req.Query.Query}
+
+		if len(req.Query.Variables) > 0 {
+			dgoReq.Vars = req.Query.Variables
+		}
+
+		resp, err := tx.Do(ctx, dgoReq)
+		if err != nil {
+			return Response{}, err
+		}
+
+		return Response{Json: string(resp.Json), Uids: resp.Uids}, nil
+	}
+
+	tx := dc.dgClient.NewTxn()
+	defer func() {
+		if err := tx.Discard(ctx); err != nil {
+			logger.Warn(ctx).Err(err).Msg("Error discarding transaction.")
+			return
+		}
+
+	}()
+
+	mus := make([]*api.Mutation, 0, len(req.Mutations))
+
+	for _, m := range req.Mutations {
+		mu := &api.Mutation{}
+		if m.SetJson != "" {
+			mu.SetJson = []byte(m.SetJson)
+		}
+		if m.DelJson != "" {
+			mu.DeleteJson = []byte(m.DelJson)
+		}
+		if m.SetNquads != "" {
+			mu.SetNquads = []byte(m.SetNquads)
+		}
+		if m.DelNquads != "" {
+			mu.DelNquads = []byte(m.DelNquads)
+		}
+		if m.Condition != "" {
+			mu.Cond = m.Condition
+		}
+
+		mus = append(mus, mu)
+	}
+
+	dgoReq := &api.Request{Mutations: mus, CommitNow: req.CommitNow}
+
+	if req.Query.Query != "" {
+		dgoReq.Query = req.Query.Query
+
+		if len(req.Query.Variables) > 0 {
+			dgoReq.Vars = req.Query.Variables
+		}
+	}
+
+	resp, err := tx.Do(ctx, dgoReq)
+	if err != nil {
+		return Response{}, err
+	}
+
+	return Response{Json: string(resp.Json), Uids: resp.Uids}, nil
 }
 
 func (dc *dgraphConnector) executeUpserts(ctx context.Context, query string, setMutations, delMutations []string) (map[string]string, error) {