+++ hostfunctions/hostfunctions.go
@@ -10,7 +10,7 @@ import (
 	"fmt"
 	"time"
 
-	"hmruntime/dqlclient"
+	"hmruntime/dgraphclient"
 	"hmruntime/languages"
 	"hmruntime/logger"
 	"hmruntime/manifestdata"
@@ -43,7 +43,7 @@ func RegisterHostFunctions(ctx context.Context) {
 
 	manifestdata.RegisterManifestLoadedCallback(func(ctx context.Context) error {
 		sqlclient.ShutdownPGPools()
-		dqlclient.ShutdownConns()
+		dgraphclient.ShutdownConns()
 		return nil
 	})
 }