+++ services/services.go
@@ -10,7 +10,7 @@ import (
 	"hmruntime/aws"
 	"hmruntime/collections"
 	"hmruntime/db"
-	"hmruntime/dqlclient"
+	"hmruntime/dgraphclient"
 	"hmruntime/graphql"
 	"hmruntime/hostfunctions"
 	"hmruntime/logger"
@@ -56,7 +56,7 @@ func Stop(ctx context.Context) {
 
 	collections.CloseIndexFactory(ctx)
 	sqlclient.ShutdownPGPools()
-	dqlclient.ShutdownConns()
+	dgraphclient.ShutdownConns()
 	wasmhost.RuntimeInstance.Close(ctx)
 	logger.Close()
 	db.Stop(ctx)