+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Improve logger registration code [#335](https://github.com/hypermodeAI/runtime/pull/335)
 - Add dgraph host functions [#336](https://github.com/hypermodeAI/runtime/pull/336)
+- Add upsert to dgraph host functions [#342](https://github.com/hypermodeAI/runtime/pull/342)
 
 ## 2024-08-12 - Version 0.11.0